Release 1.0.0 (#176)
* Add all sounds: #85 und #84 (#152)

+ Added all Weapon Sounds
+ Added UI Sounds
+ SFX and MusicSlide work
* MasterSlide does not work!

* Feature/ci cd into develop (#154)

* Updated Changelog: #45

* update Changelog: #45

* Added Test Run: #61

* Updated acitvation-yml: #61

* Added Builder for WebGl and Windows: #61

* Fixed projectPath error and Added deployPage: #61

* Fixed projectPath Error for WindowsBuild

* Testing new TestRunner: #61 

Removed all Builder and Deployments

* Updated Builders in template.yml: #61

* Added complete new TestRunner: #61

* Added new Build for windows: #61

* Forgot needing TestRunner for Build: #61

* Took example .yml to test what the problem: #61

* Added projectPath Line: #61

* Updated template.aml

* #61

* #61

* #61

* Feature/UI (#153)

* Worked on HelperUI #132 #63

* #63 #62

equipped Wepons and Items are shown, fix in items: add varity again

* Fixed Bug for Build

* bugfix in changeAccount methods in account.js and Database.cs, other minor changes (#155)

* no unnecesary database operations on changeAccount, playerTag changable

* deleted not implemented requests in test.http

* bugfix new playerTag was in url param when playerTag was changed

* Shop manager clean code (#156)

* put method logic in smaller methods in order to make them less complex

* removed not used using directive

* CI/CD into develop  (#157)

* Workflow only on pushes

* Added WebGL Builder in Unity: #61

* updated yml , starts only when merging on main: #61

* Added LinuxBuild : #61

* clean code in Item, Item balancing, minor bugfix, Character movement speed gets calculated (#158)

* put method logic in smaller methods in order to make them less complex

* removed not used using directive

* Clean code in Item, bugfix in Characters.cs

made variables in item private and added properties
added maxPsychoLevel variable

* removed not used using directives, implemented changes of Item in ShopManager

* changed Item scriptable Objects, deleted empty directories

* character movementSpeed as impact on actual movementSpeed now, bugfix in GetPercentage method in Weapon

* Bugfix in template.yml

* Boss EnemyAI Behaviour, Enemy Stat Scaling #83 (#159)

* Create Branch BossAIBehaviour

* Merge branch develop into feature/BossAIBehaviour

* #83 Unique AI behaviour for boss

-Add unique attack for RangedBoss
-Update movement pattern

* BossPrrojectile ignore Enemy collision

-Fixed BossProjectile colliding with Enemies

* #83 Add BossAI behaviour

-Add Boss Movement and Attack
-Add Scaling to all Enemy stats

* Fix Scaling

-Preset Adequate Values
-Fix Scaling Issues

* Adjust Wave Spawning

-Adjust Values for Enemies Spawning

* Feature/unit tests in develop (#162)

* Tests für alle Mods: #76

* Fixed Commit: #76

* Fix weapon bugs and clean code (#170)

* #163 Resolve Bug in Hacking, GorillaArms, Katana, Revolver

* #167 Fix Weapon inventory reset bug

Converting the singleton WeaponInventory to Monobehaviour reloads the script and its properties.

* Change Weapon property access modifiers to protected where possible

* Remove lifesteal properties from weapon classes

* Revert portected properties in weapon classes to public

Unity tests need to access these protected properties.

* Remove damage enemy test

This test was not working as intended.

* Added Doku (#171)

* Queued Spawning, Bugfixes, Cleanup (#172)

* Create Branch BossAIBehaviour

* Merge branch develop into feature/BossAIBehaviour

* #83 Unique AI behaviour for boss

-Add unique attack for RangedBoss
-Update movement pattern

* BossPrrojectile ignore Enemy collision

-Fixed BossProjectile colliding with Enemies

* #83 Add BossAI behaviour

-Add Boss Movement and Attack
-Add Scaling to all Enemy stats

* Fix Scaling

-Preset Adequate Values
-Fix Scaling Issues

* Adjust Wave Spawning

-Adjust Values for Enemies Spawning

* Queued Enemy Spawning, SpawnCap Bugfix

-Add Queued Spawning with SpawnCap
-SpawnCap exceeding Bug fixed

* Feature/UI (#173)

* add characterpicture in shop

* Add deleteAccount Button in MainMenu

* fix LoginButton und deleteAccountButton

* bugfix playerData was not saved after logout, removed debug.log

* Update MainMenu.cs

---------


* Clean code (#175)

* updated ChangeLog: no issue left

* cleancode in all classes

---------


---------

+++ WASD/Assets/Scripts/Player/PlayerHealthManager.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
@@ -13,7 +12,9 @@ public class PlayerHealthManager : MonoBehaviour
     {
         get { return currentHealth; }
     }
-    private void Start() {
+    
+    private void Start()
+    {
         currentChar = CharactersManager.CurrentChar;
         currentHealth = currentChar.MaxHealthValue;
     }
@@ -46,7 +47,7 @@ public class PlayerHealthManager : MonoBehaviour
         //lerpTimer = 0f;
     }
 
-    public void DamagePlayer(int damage)
+    public void DamagePlayer(float damage)
     {
         if((currentHealth - damage) <= 0)
         {
